:recycle: Split pipeline_handler.go
